Extracted symbol table to separate file
